% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bndes.R
\name{query_bndes}
\alias{query_bndes}
\title{Queries the loans made through The Brazilian Development Bank (BNDES) since 2002.}
\usage{
query_bndes()
}
\description{
Queries the loans made through The Brazilian Development Bank (BNDES) since 2002.
}
