% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bndespar_desimbursements.R
\name{query_bndespar_desimbursements}
\alias{query_bndespar_desimbursements}
\title{Queries BNDESPar desimbursements data}
\usage{
query_bndespar_desimbursements(year = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded. integer.}
}
\value{
a dataframe with the data
}
\description{
Downloads data from BNDESPar desimbursements since 2007 and return it in the form of a dataframe.
}
\examples{
df <- query_bndespar_desimbursements()

}
