% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bndespar_portfolio.R
\name{query_bndespar_portfolio}
\alias{query_bndespar_portfolio}
\title{Queries BNDESPar portfolio data}
\usage{
query_bndespar_portfolio(year = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded. integer.}
}
\value{
a dataframe with the data
}
\description{
Downloads data from the portfolio since 2006 and return it in the form of a dataframe.
}
\examples{
df <- query_bndespar_portfolio()

}
