% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_contracts_data.R
\name{query_contracts_data}
\alias{query_contracts_data}
\title{Query the data of the loan contracts made through the National Bank for Economic and Social Development (BNDES).}
\usage{
query_contracts_data(year = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded. integer.}
}
\value{
a dataframe with data for the selected year.
}
\description{
Downloads contracts data for the selected years, since 2002, and return it in the form of a dataframe.
Note: to access the total amount disbursed, use the function query_desimbursements_data().
}
\examples{
\dontrun{query_contracts_data(year = 2012)}

}
