% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_desimbursements.R
\name{query_desimbursements}
\alias{query_desimbursements}
\title{Query data on loan disbursements made through the National Bank for Economic and Social Development (BNDES).}
\usage{
query_desimbursements(year = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded. integer.}
}
\value{
a dataframe with data for the selected year.
}
\description{
Downloads the data on loan disbursements for the selected years, since 1995, and return it in the form of a dataframe.
}
\examples{
\dontrun{query_desimbursements(year = c(1999:2010))}

}
